+++ public_transport_watcher/db/migrations/alembic/versions/79463770f866_added_journey_id.py
@@ -11,7 +11,6 @@ from typing import Sequence, Union
 from alembic import op
 import sqlalchemy as sa
 
-
 # revision identifiers, used by Alembic.
 revision: str = "79463770f866"
 down_revision: Union[str, None] = "a8ff31fc08d1"