+++ public_transport_watcher/predictor/configuration/__init__.py
@@ -1,0 +1,3 @@
+from .prediction_config import PREDICTION_CONFIG
+
+__all__ = ["PREDICTION_CONFIG"]