+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.11.0] - 2025-05-05
+MR #12
+
+### Added
+- Graph builder, saver, and loader for the public transport network
+- Visualizer for the network
+- `networkx` in the dependencies
+- `NETWORK_PATH` in the `.env.example` file
+
 ## [0.10.0] - 2025-04-30
 MR #11